Changing tests and methods
Have made some changes to my tests and to the methods. Have made all tests become equal to true, and have changed the methods by putting a ? in from of them. Have done this so as to make the methods return true or false boolean statements, as my plan for trying to refactor this and keep the methods would require me to do this. Tests still pass, and will now try and refactor this into one method that checks all numbers.