Refactoring
Have refactored the factor.rb into one method that is called and checks all numbers that are passed into it. Instead of having the 8 different methods that all can be called, have put the initial methods I made into private and have put these into the 'check_number' method I have created so that the number will be checked that way. Have also deleted the 'number?' method as I feel this is not needed. Made the changes to my tests just to make sure that this will work, and all tests run pass their necessary tests. Going to look if there is a better way to write my method as I don't like the length of the 'if' and 'else' statements